requirements.txt
numpy
pandas
biopython
torch
tensorflow
scikit-learn
matplotlib
seaborn
pyyaml
Flask
